Re-factor a bunch of the python code
- JS side has not been refactored yet
- Have a base Element type that can be used by HTML elements, other basic elements
- FunctionElement type handles Elements that can re-render with a function
- Separate components and object_types out into their own folder
- Motivated by keeping separation of concerns. Theoretically the render/elements could all by in their own package that is not dependent on `deephaven` package at all, which may be advantageous to separate out at some point (same rendering backend with different rendering frontend)
  - Would need to move the use_table_listener hook as well
