+++ plugins/ui/src/deephaven/ui/__init__.py
@@ -3,7 +3,7 @@
 from deephaven.plugin import Registration, Callback
 from .components import *
 from .hooks import *
-
+from .object_types import *
 
 __version__ = "0.0.1.dev0"
 
@@ -11,6 +11,6 @@ __version__ = "0.0.1.dev0"
 class UIRegistration(Registration):
     @classmethod
     def register_into(cls, callback: Callback) -> None:
-        callback.register(UINodeType)
+        callback.register(ElementType)
+        callback.register(FunctionElementType)
         callback.register(TextFieldType)
-        callback.register(TextType)
