+++ plugins/ui/src/deephaven/ui/components/__init__.py
@@ -1,23 +1,15 @@
-from .Element import Element, ElementType
-from .node import UINode, UINodeType
-from .text import TextType, text
-from .text_field import TextFieldType, text_field
-from .make_component import make_component as component
 from .flex import flex, flex_col, flex_row
-import html
+from .make_component import make_component as component
+from .text import text
+from .text_field import text_field
+
 
 __all__ = [
     "component",
-    "Element",
-    "ElementType",
     "flex",
     "flex_col",
     "flex_row",
     "html",
     "text",
-    "TextType",
     "text_field",
-    "TextFieldType",
-    "UINode",
-    "UINodeType",
 ]
