+++ plugins/ui/src/deephaven/ui/components/flex.py
@@ -1,4 +1,4 @@
-from .HTMLElement import HTMLElement
+from ..elements import HTMLElement
 
 
 def flex(*children, **attributes):
