+++ plugins/ui/src/deephaven/ui/elements/__init__.py
@@ -1,0 +1,13 @@
+from .Element import Element
+from .HTMLElement import HTMLElement
+from .TextElement import TextElement
+from .TextField import TextField
+from .FunctionElement import FunctionElement
+
+__all__ = [
+    "Element",
+    "FunctionElement",
+    "HTMLElement",
+    "TextElement",
+    "TextField",
+]
