+++ plugins/ui/src/deephaven/ui/object_types/__init__.py
@@ -1,0 +1,5 @@
+from .ElementType import ElementType
+from .FunctionElementMessageStream import FunctionElementMessageStream
+from .FunctionElementType import FunctionElementType
+from .TextFieldType import TextFieldType
+from .TextFieldMessageStream import TextFieldMessageStream
